Rename rule-properties-order
