+++ docs/user-guide/example-config.md
@@ -35,6 +35,7 @@ You might want to learn a little about [the conventions](https://github.com/styl
     "declaration-block-no-duplicate-properties": true,
     "declaration-block-no-shorthand-property-overrides": true,
     "declaration-block-no-single-line": true,
+    "declaration-block-properties-order": "alphabetical"|[],
     "declaration-block-semicolon-newline-after": "always"|"always-multi-line"|"never-multi-line",
     "declaration-block-semicolon-newline-before": "always"|"always-multi-line"|"never-multi-line",
     "declaration-block-semicolon-space-after": "always"|"never"|"always-single-line"|"never-single-line",
@@ -93,7 +94,6 @@ You might want to learn a little about [the conventions](https://github.com/styl
     "root-no-standard-properties": true,
     "rule-nested-empty-line-before": "always"|"never",
     "rule-non-nested-empty-line-before": "always"|"never",
-    "rule-properties-order": "alphabetical"|[],
     "selector-class-pattern": string,
     "selector-combinator-space-after": "always"|"never",
     "selector-combinator-space-before": "always"|"never",
