+++ src/rules/declaration-block-properties-order/__tests__/index.js
@@ -1,0 +1,4 @@
+import "./alphabetical"
+import "./flat"
+import "./grouped-flexible"
+import "./grouped-strict"
