+++ src/rules/rule-properties-order/__tests__/index.js
@@ -1,4 +1,4 @@
-import "./alphabetical"
-import "./flat"
-import "./grouped-flexible"
-import "./grouped-strict"
+// import "./alphabetical"
+// import "./flat"
+// import "./grouped-flexible"
+// import "./grouped-strict"
